Fixed a bug

+++ test_/test.c
@@ -1,12 +1,12 @@
 static double ale;
 
-int pass_parameters_3(double d1, double d2, int i1, double d3, double d4,
+double pass_parameters_3(double d1, double d2, int i1, double d3, double d4,
  double d5, double d6, unsigned int i2, long i3,
  double d7, double d8, unsigned long i4, double d9,
  int i5, double d10, int i6, int i7, double d11,
  int i8, int i9)
  {
-    return 1;
+    return 1.0;
  }
